+++ src/components/Board.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { makeMove } from '../actions';
+import { makeMove } from '../actions/game';
 import { Card } from './Card';
 
 const BoardPure = ({ board, makeMove }) => {
